third_party/WebKit/LayoutTests/inspector-protocol/debugger/scriptParsedHash.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>
function test()
{
    var hashes = new Set(["1C6D2E82E4E4F1BA4CB5762843D429DC872EBA18",
                          "EBF1ECD351E7A3294CB5762843D429DC872EBA18",
                          "22D0043331237371241FC675A984B967025A3DC0"]);
    InspectorTest.sendCommandOrDie("Debugger.enable", {}, function() {
        InspectorTest.eventHandler["Debugger.scriptParsed"] = function(messageObject)
        {
            if (hashes.has(messageObject.params.hash))
                InspectorTest.log("Hash received: " + messageObject.params.hash);
        }
    });

    function longScript() {
        var longScript = "var b = 1;";
        for (var i = 0; i < 2024; ++i)
            longScript += "++b;";
    }

    InspectorTest.sendCommandOrDie("Runtime.evaluate", { expression: "1" });
    InspectorTest.sendCommandOrDie("Runtime.evaluate", { expression: "239" });
    InspectorTest.sendCommandOrDie("Runtime.evaluate", { expression: "(" + longScript + ")()" }, step2);

    function step2()
    {
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onLoad="runTest();">
</body>
</html>
